Move websocket to own module
